requirements.txt
langgraph
langchain
openai
python-dotenv
pydantic
rich
typer